refactor(cmd): rename kill => process
This commit contains cosmetic changes only.

Rename the following:

* kong.cmd.utils.kill              => kong.cmd.utils.process
* kong.cmd.utils.kill.is_running() => kong.cmd.utils.process.exists()

The purpose for this is really just ergonomics. The name `kill` was
obviously inherited from the syscall by the same name, and while I don't
really think it's a "bad" word to use, it's also not a very accurate
description of what the module is for.